Initial upload of custom training loop over sequence segments
Loads the data, initialises the parameters, initialises the optimiser,
model and regularizer.
During the training loop it iterates over the segments of each sequence.
Once a sequence is finished, it iterates over the validation data.

After the training loop ends, it plots the metrics for train, validation
and test data